+++ organizations/contentstrategies/organizations_content_strategies_test/organizations_content_strategies_test.go
@@ -93,6 +93,11 @@ func TestOrganizationsContentStrategiesCreateWithWireMock(
 			Attributes: &organizations.CreateContentStrategyAttributes{
 				Name: "Featured Travel",
 				Sort: organizations.ContentStrategySortHighestCashback.Ptr(),
+				Filters: &organizations.ContentStrategyFilters{
+					OfferFeatures: []organizations.OfferFeatures{
+						organizations.OfferFeaturesInteractive,
+					},
+				},
 				Categories: []kard.CategoryOption{
 					kard.CategoryOptionTravel,
 				},
@@ -178,7 +183,8 @@ func TestOrganizationsContentStrategiesUpdateWithWireMock(
 	request := &organizations.UpdateContentStrategyRequestBody{
 		Data: &organizations.UpdateContentStrategyRequestData{
 			Attributes: &organizations.UpdateContentStrategyAttributes{
-				Name: "name",
+				Name:    "name",
+				Filters: &organizations.ContentStrategyFilters{},
 				Categories: []kard.CategoryOption{
 					kard.CategoryOptionArtsEntertainment,
 					kard.CategoryOptionArtsEntertainment,